tests/wpt/tests/permissions-policy/experimental-features/permissions-policy-header-scheme-only.https.sub.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script>
  test(function() {
    var wildcard_origin = 'https:';
    assert_array_equals(
      document.featurePolicy.getAllowlistForFeature('fullscreen'),
      [wildcard_origin]);
    assert_true(document.featurePolicy.allowsFeature('fullscreen'));
  }, 'Test scheme only policy works as expected');
  </script>
</body>
